fix cut off documents caused by skipLine and code sections
I posted a diff into one of my posts within a fenced code section.
As soon as convert_gopher and convert_gemini encounterned a line
with just ---, the document would be cut off in the line after.

Such lines can be encountered for example in a Unix mailbox formatted
email.